Split check/lints.rs into themed lints/ submodules
Pure move-only refactor: the 3,471-line check/lints.rs monolith becomes a
directory module `check/lints/` with a facade `mod.rs` plus six themed
submodules. No behavior or generated-output changes.

- lints/mod.rs       — `check_completeness` orchestrator + submodule
                       declarations and `pub(crate) use <m>::*` re-exports,
                       so both `crate::check::<sym>` (via check/mod.rs's
                       `pub use lints::*`) and `crate::check::lints::<sym>`
                       keep resolving.
- lints/shared.rs    — contains_word, parse_top_level_cmp, classify_field_type
                       + FieldTypeShape, make_old_in_single_state_warning,
                       ref_impl_has_overflow_risk.
- lints/arithmetic.rs— ref_impl unbounded arith, checked-arith math-overflow,
                       wrapping opt-in, Map/subscript validation.
- lints/cpi.rs       — pre/post field extraction, multi-CPI shared-field,
                       no-callee-ensures, unverified-callee, shape-only CPI,
                       collect_require_verified_findings + UnverifiedCallee.
- lints/state.rs     — old-in-single-state, unconstrained modifies, terminal/
                       indexed/dedup transition gaps, cross-ADT ambiguity.
- lints/structural.rs— error-as-record, unknown error variant, PDA collisions,
                       vacuous property lowering.
- lints/auth.rs      — r25_will_bind_auth predicate, unbound_auth,
                       unconditional_value_transfer.

Visibility is unchanged (every pub(crate) helper stays pub(crate)) so
check/tests.rs keeps reaching the helpers through the facade. Gate: cargo
build + cargo test (1006 unit + all snapshot suites) + cargo fmt --check +
clippy -D warnings all pass.
